pyproject.toml
[project]
name = "llm_mem"
version = "0.1.0"
description = ""
authors = [
    {name = "Preslav",email = "preslav.aleksandrov@protonmail.com"}
]
readme = "README.md"
requires-python = "3.11.9"
dependencies = [
    "flwr[simulation]>=1.14.0",
    "flwr-datasets[vision]>=0.3.0",
    "scipy (>=1.15.0,<2.0.0)",
    "numpy (>=2.2.1,<3.0.0)",
    "transformers (>=4.47.1,<5.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "trl (>=0.13.0,<0.14.0)",
    "torch (>=2.5.1,<3.0.0)",
    "torch @ https://download.pytorch.org/whl/cu124/torch-2.5.1%2Bcu124-cp311-cp311-linux_x86_64.whl",
    "datasets (==3.1.0)",
    "omegaconf (>=2.3.0,<3.0.0)",
    "peft (>=0.14.0,<0.15.0)",
    "evaluate (>=0.4.3,<0.5.0)",
    "wandb (>=0.19.1,<0.20.0)",
    "rouge-score (>=0.1.2,<0.2.0)",
    "absl-py (>=2.1.0,<3.0.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "onnxruntime (==1.20.1)",
    "torchvision (==0.20.1)",
    "optimum (==1.23.3)",
    "huggingface-hub[cli] (>=0.27.1,<0.28.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "boto3 (>=1.36.8,<2.0.0)",
    
]
[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
